🔀 :: stock 엔티티 추가

+++ src/main/java/com/juicycool/backend/domain/day/Day.java
@@ -36,7 +36,7 @@ public class Day {
 
     private Long volume; // 거래량
 
-    private Long upDownPercent; // 등략률
+    private Double upDownPercent; // 등략률
 
     private String storeAt; // 저장된 시간
 }
