+++ src/main/java/com/juicycool/backend/domain/day/presentation/dto/response/GetDayResponseDto.java
@@ -12,6 +12,6 @@ public class GetDayResponseDto {
     private Long presentPrice; // 현재가
     private Long headPrice; // 종가
     private Long volume; // 거래량
-    private Long upDownPercent; // 등략률
+    private Double upDownPercent; // 등략률
     private String storeAt; // 저장된 시간
 }
